requirements.txt
slack-bolt
openai
python-dotenv
flask
requests
slack_sdk
